Scaling relative to points
Summary: window->contentView is now in points. All sizes should now be relative to points.

Test Plan: Move mac app window between different DPI monitors, scaling should be the same. Resize the window, scaling should be the same.

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: mike.naquin

Differential Revision: https://phabricator.watchonair.tv/D980

+++ okui/include/onair/okui/views/Button.h
@@ -27,8 +27,8 @@ public:
     virtual void render() override;
 
     virtual bool canBecomeFocus() override { return true; }
-    virtual void mouseDown(MouseButton button, int x, int y) override;
-    virtual void mouseUp(MouseButton button, int startX, int startY, int x, int y) override;
+    virtual void mouseDown(MouseButton button, double x, double y) override;
+    virtual void mouseUp(MouseButton button, double startX, double startY, double x, double y) override;
     virtual void mouseExit() override;
     virtual void keyDown(KeyCode key, KeyModifiers mod, bool repeat) override;
     virtual void windowChanged() override;
