+++ okui/include/onair/okui/views/TextField.h
@@ -75,8 +75,8 @@ public:
     virtual void focusGained() override;
     virtual void focusLost() override;
 
-    virtual void mouseDown(MouseButton button, int x, int y) override;
-    virtual void mouseDrag(int startX, int startY, int x, int y) override;
+    virtual void mouseDown(MouseButton button, double x, double y) override;
+    virtual void mouseDrag(double startX, double startY, double x, double y) override;
     virtual void mouseEnter() override;
     virtual void mouseExit() override;
     virtual void textInput(const std::string& text) override;
